# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ini adalah dokumentasi API untuk sistem pencatatan target dan performa Karyawan (Employee Tracker).",
        "title": "Employee Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/achievements": {
            "post": {
                "description": "Mencatat riwayat transaksi/pencairan (ledger) baru untuk sebuah target",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "Catat Pencapaian Target",
                "parameters": [
                    {
                        "description": "Payload data pencapaian",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.RecordAchievementRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pencapaian berhasil dicatat",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_AchievementResponse"
                        }
                    },
                    "400": {
                        "description": "Format JSON salah atau validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Target tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/employees": {
            "get": {
                "description": "Menampilkan semua data karyawan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Data Karyawan",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail karyawan",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-array_github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_EmployeeResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Mendaftarkan data karyawan baru ke dalam sistem",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Registrasi Karyawan Baru",
                "parameters": [
                    {
                        "description": "Payload data karyawan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.EmployeeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Karyawan berhasil didaftarkan",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_EmployeeResponse"
                        }
                    },
                    "400": {
                        "description": "Format JSON salah atau validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal menyimpan data karyawan",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/employees/{id}": {
            "get": {
                "description": "Menampilkan detail informasi karyawan berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Detail Karyawan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Karyawan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail karyawan",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_EmployeeResponse"
                        }
                    },
                    "400": {
                        "description": "Format UUID salah",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Karyawan tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Memperbaharui data karyawan yanga ada di dalam sistem",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Update Karyawan Baru",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Karyawan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload data karyawan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.EmployeeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Karyawan berhasil diperbaharui",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_EmployeeResponse"
                        }
                    },
                    "400": {
                        "description": "Format JSON salah atau validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Karyawan tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal menyimpan data karyawan",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/targets/{target_id}/achievements": {
            "get": {
                "description": "Mengambil riwayat/ledger pencapaian untuk satu target spesifik",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "List Pencapaian Berdasarkan Target",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Target",
                        "name": "target_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil riwayat pencapaian",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-array_github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_AchievementResponse"
                        }
                    },
                    "400": {
                        "description": "Format UUID salah",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Target tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.AchievementResponse": {
            "type": "object",
            "properties": {
                "closing_date": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nominal": {
                    "type": "integer"
                },
                "target_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.EmployeeRequest": {
            "type": "object",
            "required": [
                "entry_date",
                "name",
                "office_location",
                "position"
            ],
            "properties": {
                "entry_date": {
                    "type": "string",
                    "example": "2024-01-15T08:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "Budi Santoso"
                },
                "office_location": {
                    "type": "string",
                    "example": "Cabang Utama Jakarta"
                },
                "position": {
                    "type": "string",
                    "example": "Sales Executive"
                }
            }
        },
        "github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.EmployeeResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "entry_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "office_location": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.RecordAchievementRequest": {
            "type": "object",
            "required": [
                "nominal",
                "target_id"
            ],
            "properties": {
                "closing_date": {
                    "type": "string",
                    "example": "2026-05-10T14:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Pencairan KUR Nasabah A"
                },
                "nominal": {
                    "type": "integer",
                    "example": 150000000
                },
                "target_id": {
                    "type": "string",
                    "example": "uuid-target"
                }
            }
        },
...
        "github_com_faridlan_employee-tracker-api_internal_utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error": {
                    "type": "string",
                    "example": "pesan error"
                }
            }
        },
        "github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-array_github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_AchievementResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.AchievementResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-array_github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_EmployeeResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.EmployeeResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_AchievementResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.AchievementResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faridlan_employee-tracker-api_internal_utils.SuccessResponse-github_com_faridlan_employee-tracker-api_internal_delivery_http_dto_EmployeeResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_faridlan_employee-tracker-api_internal_delivery_http_dto.EmployeeResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}